トークンの値を取得する
